Refactor Root folder structure
