+++ src/UI/Pages/PrimaryPageManager.tsx
@@ -1,23 +1,23 @@
 import React from "react";
 import { PageManager, Page, RouteDictionary } from "@/Core";
-import { AgentProcessPage } from "@/UI/Pages/AgentProcess";
-import { AgentsPage } from "@/UI/Pages/Agents";
-import { CompileDetailsPage } from "@/UI/Pages/CompileDetails";
-import { CompileReportsPage } from "@/UI/Pages/CompileReports";
-import { CreateEnvironmentPage } from "@/UI/Pages/CreateEnvironment";
-import { CreateInstancePage } from "@/UI/Pages/CreateInstance";
-import { DesiredStatePage } from "@/UI/Pages/DesiredState";
-import { DiagnosePage } from "@/UI/Pages/Diagnose";
-import { EditInstancePage } from "@/UI/Pages/EditInstance";
-import { EventsPage } from "@/UI/Pages/Events";
-import { HomePage } from "@/UI/Pages/Home";
-import { ResourceDetailsPage } from "@/UI/Pages/ResourceDetails";
-import { ResourcesPage } from "@/UI/Pages/Resources";
-import { ServiceCatalogPage } from "@/UI/Pages/ServiceCatalog";
-import { ServiceInstanceHistoryPage } from "@/UI/Pages/ServiceInstanceHistory";
-import { ServiceInventoryPage } from "@/UI/Pages/ServiceInventory";
-import { SettingsPage } from "@/UI/Pages/Settings";
-import { StatusPage } from "@/UI/Pages/Status";
+import { AgentProcessPage } from "./AgentProcess";
+import { AgentsPage } from "./Agents";
+import { CompileDetailsPage } from "./CompileDetails";
+import { CompileReportsPage } from "./CompileReports";
+import { CreateEnvironmentPage } from "./CreateEnvironment";
+import { CreateInstancePage } from "./CreateInstance";
+import { DesiredStatePage } from "./DesiredState";
+import { DiagnosePage } from "./Diagnose";
+import { EditInstancePage } from "./EditInstance";
+import { EventsPage } from "./Events";
+import { HomePage } from "./Home";
+import { ResourceDetailsPage } from "./ResourceDetails";
+import { ResourcesPage } from "./Resources";
+import { ServiceCatalogPage } from "./ServiceCatalog";
+import { ServiceInstanceHistoryPage } from "./ServiceInstanceHistory";
+import { ServiceInventoryPage } from "./ServiceInventory";
+import { SettingsPage } from "./Settings";
+import { StatusPage } from "./Status";
 
 export class PrimaryPageManager implements PageManager {
   constructor(private readonly routeDictionary: RouteDictionary) {}
