+++ src/UI/Pages/index.ts
@@ -13,3 +13,4 @@ export * from "./Home";
 export * from "./CreateEnvironment";
 export * from "./ResourceDetails";
 export * from "./Status";
+export * from "./PrimaryPageManager";
