--- src/UI/Root/AppLayout/Toolbar/index.ts
@@ -1,5 +1,0 @@
-export * from "./Profile";
-export * from "./DocumentationLink";
-export * from "./SettingsButton";
-export * from "./EnvSelector";
-export * from "./StatusButton";
