+++ src/UI/Root/Components/AuthProvider/index.ts
@@ -1,0 +1,1 @@
+export * from "./AuthProvider";
