+++ src/UI/Root/Components/Initializer/index.ts
@@ -1,0 +1,1 @@
+export * from "./Initializer";
