+++ src/UI/Root/Components/PageBreadcrumbs/index.ts
@@ -1,0 +1,1 @@
+export * from "./PageBreadcrumbs";
