+++ src/UI/Root/Components/PageFrame/index.ts
@@ -1,0 +1,1 @@
+export * from "./PageFrame";
