+++ src/UI/Root/Components/Sidebar/EnvironmentControls/HaltDialog.tsx
