+++ src/UI/Root/Components/Sidebar/index.ts
@@ -1,0 +1,1 @@
+export * from "./Sidebar";
