+++ src/UI/Root/Root.tsx
@@ -1,13 +1,9 @@
 import React, { useContext } from "react";
-import { KeycloakProvider } from "react-keycloak";
 import { Navigate, Route, Routes } from "react-router-dom";
-import { Spinner, Bullseye } from "@patternfly/react-core";
 import { DependencyContext } from "@/UI/Dependency";
-import { NotFoundPage } from "@/UI/Pages";
+import { NotFoundPage, PrimaryPageManager } from "@/UI/Pages";
 import { SearchSanitizer } from "@/UI/Routing";
-import { AppFrame } from "./AppFrame";
-import { Initializer } from "./Initializer";
-import { PrimaryPageManager } from "./PrimaryPageManager";
+import { AuthProvider, PageFrame, Initializer } from "./Components";
 
 export const Root: React.FC = () => {
   const { routeManager } = useContext(DependencyContext);
@@ -18,7 +14,7 @@ export const Root: React.FC = () => {
   return (
     <Initializer>
       <SearchSanitizer.Provider>
-        <AuthWrapper>
+        <AuthProvider>
           <Routes>
             {routeManager.isBaseUrlDefined() && (
               <Route
@@ -31,45 +27,25 @@ export const Root: React.FC = () => {
             <Route
               path="*"
               element={
-                <AppFrame environmentRole="Optional">
+                <PageFrame environmentRole="Optional">
                   <NotFoundPage />
-                </AppFrame>
+                </PageFrame>
               }
             />
             {pages.map(({ path, kind, element, environmentRole }) => (
               <Route
                 path={path}
                 element={
-                  <AppFrame environmentRole={environmentRole}>
+                  <PageFrame environmentRole={environmentRole}>
                     {element}
-                  </AppFrame>
+                  </PageFrame>
                 }
                 key={kind}
               />
             ))}
           </Routes>
-        </AuthWrapper>
+        </AuthProvider>
       </SearchSanitizer.Provider>
     </Initializer>
-  );
-};
-
-const AuthWrapper: React.FC = ({ children }) => {
-  const { keycloakController } = useContext(DependencyContext);
-
-  if (!keycloakController.isEnabled()) return <>{children}</>;
-
-  return (
-    <KeycloakProvider
-      keycloak={keycloakController.getInstance()}
-      initConfig={keycloakController.getInitConfig()}
-      LoadingComponent={
-        <Bullseye>
-          <Spinner size="xl" />
-        </Bullseye>
-      }
-    >
-      {children}
-    </KeycloakProvider>
   );
 };
